requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
python-docx>=0.8.11
lxml>=4.6.0
